# Conflicts:
#	BE_Project/src/main/java/com/example/be_project/model/Customer.java
#	BE_Project/src/main/java/com/example/be_project/model/CustomerAddress.java

+++ BE_Project/src/main/java/com/example/be_project/model/Message.java
@@ -22,7 +22,7 @@ public class Message {
     private Date messageDatetime;
 
     @Column(name = "seller_id")
-    private Long sellerId;
+    private int sellerId;
 
     // Getter and setter methods
 }
